requirements.txt
httpx
fastapi
uvicorn
python-dotenv
sqlalchemy
alembic
asyncpg
pydantic
aiogram
psycopg2-binary